+++ maitake/src/wait/semaphore.rs
@@ -25,7 +25,9 @@ use mycelium_util::fmt;
 use mycelium_util::sync::CachePadded;
 use pin_project::{pin_project, pinned_drop};
 
-#[cfg(test)]
+#[cfg(all(test, loom))]
+mod loom;
+#[cfg(all(test, not(loom)))]
 mod tests;
 
 #[derive(Debug)]
