+++ maitake/src/wait/semaphore/tests.rs
@@ -1,174 +1,5 @@
 use super::*;
 
-#[cfg(loom)]
-mod loom {
-    use super::*;
-    use crate::loom::{
-        self, future,
-        sync::{
-            atomic::{AtomicUsize, Ordering::SeqCst},
-            Arc,
-        },
-        thread,
-    };
-
-    #[test]
-    fn basically_works() {
-        const TASKS: usize = 2;
-
-        async fn task((ref sem, ref count): &(Semaphore, AtomicUsize)) {
-            let permit = sem.acquire(1).await.unwrap();
-            let actual = count.fetch_add(1, SeqCst);
-            assert!(actual <= TASKS - 1);
-
-            let actual = count.fetch_sub(1, SeqCst);
-            assert!(actual <= TASKS);
-            drop(permit);
-        }
-
-        loom::model(|| {
-            let sem = Arc::new((Semaphore::new(TASKS), AtomicUsize::new(0)));
-            let threads = (0..TASKS)
-                .map(|_| {
-                    let sem = sem.clone();
-                    thread::spawn(move || {
-                        future::block_on(task(&sem));
-                    })
-                })
-                .collect::<Vec<_>>();
-
-            future::block_on(task(&sem));
-
-            for t in threads {
-                t.join().unwrap();
-            }
-        })
-    }
-
-    #[test]
-    fn release_on_drop() {
-        loom::model(|| {
-            let sem = Arc::new(Semaphore::new(1));
-
-            let thread = thread::spawn({
-                let sem = sem.clone();
-                move || {
-                    let _permit = future::block_on(sem.acquire(1)).unwrap();
-                }
-            });
-
-            let permit = future::block_on(sem.acquire(1)).unwrap();
-            drop(permit);
-            thread.join().unwrap();
-        })
-    }
-
-    #[test]
-    fn close() {
-        loom::model(|| {
-            let sem = Arc::new(Semaphore::new(1));
-            let threads: Vec<_> = (0..2)
-                .map(|_| {
-                    thread::spawn({
-                        let sem = sem.clone();
-                        move || -> Result<(), ()> {
-                            for _ in 0..2 {
-                                let _permit = future::block_on(sem.acquire(1)).map_err(|_| ())?;
-                            }
-                            Ok(())
-                        }
-                    })
-                })
-                .collect();
-
-            sem.close();
-
-            for thread in threads {
-                let _ = thread.join().unwrap();
-            }
-        })
-    }
-
-    #[test]
-    fn concurrent_close() {
-        fn run(sem: Arc<Semaphore>) -> impl FnOnce() -> Result<(), ()> {
-            move || {
-                let permit = future::block_on(sem.acquire(1)).map_err(|_| ())?;
-                drop(permit);
-                sem.close();
-                Ok(())
-            }
-        }
-
-        loom::model(|| {
-            let sem = Arc::new(Semaphore::new(1));
-            let threads: Vec<_> = (0..2).map(|_| thread::spawn(run(sem.clone()))).collect();
-            let _ = run(sem)();
-
-            for thread in threads {
-                let _ = thread.join().unwrap();
-            }
-        })
-    }
-
-    #[test]
-    fn concurrent_cancel() {
-        use futures_util::future::FutureExt;
-        fn run(sem: &Arc<Semaphore>) -> impl FnOnce() {
-            let sem = sem.clone();
-            move || {
-                future::block_on(async move {
-                    // poll two `acquire` futures immediately and then cancel
-                    // them, regardless of whether or not they complete.
-                    let _permit1 = {
-                        let acquire = sem.acquire(1);
-                        acquire.now_or_never()
-                    };
-                    let _permit2 = {
-                        let acquire = sem.acquire(1);
-                        acquire.now_or_never()
-                    };
-                })
-            }
-        }
-
-        loom::model(|| {
-            let sem = Arc::new(Semaphore::new(0));
-
-            let thread1 = thread::spawn(run(&sem));
-            let thread2 = thread::spawn(run(&sem));
-            let thread3 = thread::spawn(run(&sem));
-
-            thread1.join().unwrap();
-            sem.add_permits(10);
-            thread2.join().unwrap();
-            thread3.join().unwrap();
-        })
-    }
-
-    #[test]
-    fn drop_permits_while_acquiring() {
-        loom::model(|| {
-            let sem = Arc::new(Semaphore::new(4));
-            let permit1 = sem
-                .try_acquire(3)
-                .expect("semaphore has 4 permits, so we should acquire 3");
-            let thread1 = thread::spawn({
-                let sem = sem.clone();
-                move || {
-                    let _permit = future::block_on(sem.acquire(2)).unwrap();
-                    assert_eq!(sem.available_permits(), 2);
-                }
-            });
-
-            drop(permit1);
-            trace!("dropped permit 1");
-            thread1.join().unwrap();
-            assert_eq!(sem.available_permits(), 4);
-        })
-    }
-}
-
 fn assert_send_sync<T: Send + Sync>() {}
 
 #[test]
@@ -189,6 +20,9 @@ fn acquire_is_send_and_sync() {
 #[cfg(feature = "alloc")]
 mod alloc {
     use super::*;
+    use crate::scheduler::Scheduler;
+    use ::alloc::sync::Arc;
+    use core::sync::atomic::AtomicBool;
 
     #[test]
     fn owned_permit_is_send_and_sync() {
@@ -198,5 +32,96 @@ mod alloc {
     #[test]
     fn acquire_owned_is_send_and_sync() {
         assert_send_sync::<AcquireOwned>();
+    }
+
+    #[test]
+    fn basic_concurrency_limit() {
+        const TASKS: usize = 8;
+        const CONCURRENCY_LIMIT: usize = 4;
+        crate::util::trace_init();
+
+        let scheduler = Scheduler::new();
+        let semaphore = Arc::new(Semaphore::new(CONCURRENCY_LIMIT));
+        let running = Arc::new(AtomicUsize::new(0));
+        let completed = Arc::new(AtomicUsize::new(0));
+
+        for _ in 0..TASKS {
+            let semaphore = semaphore.clone();
+            let running = running.clone();
+            let completed = completed.clone();
+            scheduler.spawn(async move {
+                let permit = semaphore
+                    .acquire(1)
+                    .await
+                    .expect("semaphore will not be closed");
+                assert!(test_dbg!(running.fetch_add(1, Relaxed)) < CONCURRENCY_LIMIT);
+
+                crate::future::yield_now().await;
+                drop(permit);
+
+                assert!(test_dbg!(running.fetch_sub(1, Relaxed)) <= CONCURRENCY_LIMIT);
+                completed.fetch_add(1, Relaxed);
+            });
+        }
+
+        while completed.load(Relaxed) < TASKS {
+            scheduler.tick();
+            assert!(test_dbg!(running.load(Relaxed)) <= CONCURRENCY_LIMIT);
+        }
+    }
+
+    #[test]
+    fn countdown() {
+        const TASKS: usize = 4;
+
+        let scheduler = Scheduler::new();
+        let semaphore = Arc::new(Semaphore::new(0));
+        let a_done = Arc::new(AtomicUsize::new(0));
+        let b_done = Arc::new(AtomicBool::new(false));
+
+        scheduler.spawn({
+            let semaphore = semaphore.clone();
+            let b_done = b_done.clone();
+            let a_done = a_done.clone();
+            async move {
+                tracing_02::info!("Task B starting...");
+
+                // Since the semaphore is created with 0 permits, this will
+                // wait until all 4 "A" tasks have completed.
+                let _permit = semaphore
+                    .acquire(TASKS)
+                    .await
+                    .expect("semaphore will not be closed");
+                assert_eq!(a_done.load(Relaxed), TASKS);
+
+                // ... do some work ...
+
+                tracing_02::info!("Task B done!");
+                b_done.store(true, Relaxed);
+            }
+        });
+
+        for i in 0..TASKS {
+            let semaphore = semaphore.clone();
+            let a_done = a_done.clone();
+            scheduler.spawn(async move {
+                tracing_02::info!("Task A {i} starting...");
+
+                crate::future::yield_now().await;
+
+                a_done.fetch_add(1, Relaxed);
+                semaphore.add_permits(1);
+
+                // ... do some work ...
+                tracing_02::info!("Task A {i} done");
+            });
+        }
+
+        while !b_done.load(Relaxed) {
+            scheduler.tick();
+        }
+
+        assert_eq!(a_done.load(Relaxed), TASKS);
+        assert!(b_done.load(Relaxed));
     }
 }
